+++ R/AssessGenes.R
@@ -19,7 +19,7 @@
 #'             limConCovRatio_NotCon = 0.8,
 #'             minConCovRatio_Stop = 0.5,
 #'             noConStopsGeneFrac = 0.5,
-#'             minNumStops = 2,
+#'             minNumStops = 1,
 #'             minMissORFLen = 0,
 #'             allowNestedORFs = FALSE,
 #'             useNTermProt = FALSE,
@@ -106,7 +106,7 @@
 #' The maximum of either \code{minCovNum} or (\code{minCovPct}/100) multiplied by the number of related genomes is used as
 #' the minimum coverage required in determining conserved starts and stops.
 #' 
-#' Additionaly, open reading frames with proteomics evidence but no gene start are categorized based on whether or not there
+#' Additionally, open reading frames with proteomics evidence but no gene start are categorized based on whether or not there
 #' is a conserved start upstream of the proteomic evidence. The positions and lengths of these open reading frames is included
 #' in the \code{N_CS-_PE+_ORFs} and \code{N_CS+_PE+_ORFs} matrices within the final object that is returned.
 #'
@@ -158,7 +158,7 @@ AssessGenes <- function(geneLeftPos,
                         limConCovRatio_NotCon = 0.8,
                         minConCovRatio_Stop = 0.5,
                         noConStopsGeneFrac = 0.5,
-                        minNumStops = 2L,
+                        minNumStops = 1L,
                         minMissORFLen = 0L,
                         allowNestedORFs = FALSE,
                         useNTermProt = FALSE,
@@ -369,6 +369,8 @@ AssessGenes <- function(geneLeftPos,
   
   ## --------------------------------------------------------------------------------------------------------------- ##
   
+  ## Pull the necessary vectors and lists from the mapping object.
+  
   stops <- mapObj$StopsByFrame
   
   fwdProt <- mapObj$FwdProtHits
@@ -389,36 +391,49 @@ AssessGenes <- function(geneLeftPos,
   
   ## --------------------------------------------------------------------------------------------------------------- ##
   
+  ## Check if there are enough related genomes to use conservation evidence when categorizing genes.
   if ((useCons) && (minCovNum > mapObj$NumRelatedGenomes)) {
     useCons <- FALSE
     
-    warning("The value for the minimum number of related genomes required is greater than the number of related",
-            " genomes used to generate the map object. Conserved starts & stops will not be used in assessment.")
+    warning("The value for the minimum number of related genomes required is greater than the number of related ",
+            "genomes used to generate the mapping object. Conserved starts & stops will not be used in assessment.")
   }
   
+  ## If conservation evidence can be used, determine the minimum coverage, using either the minimum
+  ## number of related genomes required or the minimum percentage of related genomes required.
   if (useCons) {
     minCov <- max(minCovNum, minCovPct * mapObj$NumRelatedGenomes / 100)
   }
   
   ## --------------------------------------------------------------------------------------------------------------- ##
   
-  checkRange <- function(x) {
-    if (all(x < 0))
-      stop("X-axis out of bounds.")
-    if (all(x > genomeLength))
-      stop("X-axis out of bounds.")
-    if (x[1] < 0)
-      x[1] <- 0
-    if (x[2] > genomeLength)
-      x[2] <- genomeLength
-    x <- floor(x)
-    return(x[1]:x[2])
+  ## If the predicted stop is downstream of the current in-frame region-ending
+  ## stop, this function will check which of the next stop-to-stop regions need
+  ## to be skipped, making sure no predicted genes will be skipped in the
+  ## process. Returns the indices of the region-ending stops to skip.
+  checkNextStops <- function(currPredStop) {
+    skipIdxs <- which((frameStops > regStart) & (frameStops < currPredStop))
+    skipIdxs <- c(skipIdxs, skipIdxs[length(skipIdxs)] + 1L)
+    
+    lastStop <- frameStops[skipIdxs[length(skipIdxs)]]
+    
+    skippedStarts <- strandFrameStarts[((strandFrameStarts > regEnd) &
+                                          (strandFrameStarts < lastStop))]
+    
+    if (length(skippedStarts) > 0) {
+      firstNoSkip <- (which(frameStops[skipIdxs] > skippedStarts[1]))[1]
+      
+      skipIdxs <- skipIdxs[seq(from = 1, to = firstNoSkip - 1, by = 1)]
+    }
+    
+    return(skipIdxs)
   }
   
-  ## This function checks if there is overlapping protein evidence in other frames. It is used
-  ## when recording information for categories with protein evidence and no gene start.
+  ## This function checks if there is overlapping protein evidence in other
+  ## frames. It is used when recording information for categories with protein
+  ## evidence and no gene start.
   checkOtherFrames <- function() {
-    currProtHitPos <- orfRange[regionalProteinHits]
+    currProtHitPos <- regRange[regionalProteinHits]
     
     sameRange <- (min(currProtHitPos)):(max(currProtHitPos))
     
@@ -436,7 +451,7 @@ AssessGenes <- function(geneLeftPos,
         if (oFrame <= 3) {
           oppStrandProt <- fwdProt[[oFrame]][[1]][sameRange]
         } else {
-          oppStrandProt <- revProt[[oFrame - 3]][[1]][orfRange]
+          oppStrandProt <- revProt[[oFrame - 3]][[1]][regRange]
         }
       }
       
@@ -448,10 +463,11 @@ AssessGenes <- function(geneLeftPos,
     return(0L)
   }
   
-  ## This function checks if the current ORF is completely within another ORF. It is used
-  ## to filter out some of the ORFs with protein hits and no given start.
+  ## This function checks if the current ORF is completely within another ORF.
+  ## It is used to filter out some of the ORFs with protein hits and no given
+  ## start.
   isNestedORF <- function() {
-    currProtHitPos <- orfRange[regionalProteinHits]
+    currProtHitPos <- regRange[regionalProteinHits]
     
     sameLeftPos <- min(currProtHitPos)
     sameRightPos <- max(currProtHitPos)
@@ -466,7 +482,7 @@ AssessGenes <- function(geneLeftPos,
         nested <- any((stops[[oFrame]][-length(stops[[oFrame]])] <= sameLeftPos) &
                         (stops[[oFrame]][-1] >= sameRightPos))
       } else {
-        ## USe opposite
+        ## Use opposite
         nested <- any((stops[[oFrame]][-length(stops[[oFrame]])] <= oppLeftPos) &
                         (stops[[oFrame]][-1] >= oppRightPos))
       }
@@ -479,26 +495,32 @@ AssessGenes <- function(geneLeftPos,
     return(FALSE)
   }
   
-  cat1A_Counter <- 0L
-  
+  ## This matrix stores information on ORFs with protein hits but no strong conserved start
+  ## upstream of the protein hits and no predicted gene
   cat1A_ORFs <- matrix(0, nrow = sum(vapply(stops, length, integer(1))), ncol = 5,
                        dimnames = list(NULL,
                                        c("Start", "End", "Length", "Frame", "OtherProtFrame")))
+  cat1A_Counter <- 0L
   
-  cat1B_Counter <- 0L
-  
+  ## This matrix stores information on ORFs with protein hits and at least one strong conserved
+  ## start upstream of the protein hits but no predicted gene.
   cat1B_ORFs <- matrix(0, nrow = sum(vapply(stops, length, integer(1))), ncol = 5,
                        dimnames = list(NULL,
                                        c("Start", "End", "Length", "Frame", "OtherProtFrame")))
+  cat1B_Counter <- 0L
   
+  ## This vector stores the categorization for each predicted gene.
   catGeneAssignment <- character(length(geneLeftPos))
   
   if (verbose) {
     pBar <- txtProgressBar(style=ifelse(interactive(), 3, 1))
   }
   
+  ## Iterate through each of the 6 frames of the genome, using
+  ## the same set of vectors to work through each frame.
   for (frameID in seq(1, 6)) {
     if (frameID <= 3) {
+      ## Get and set up the vectors to work through the current forward frame.
       strandID <- "+"
       cFrame <- frameID
       shift <- cFrame %% 3
@@ -512,11 +534,9 @@ AssessGenes <- function(geneLeftPos,
       strandOnly <- which(geneStrand == strandID)
       
       strandStarts <- geneLeftPos[strandOnly]
-      names(strandStarts) <- as.character(strandOnly)
-      
-      strandFrameStarts <- strandStarts[(strandStarts %% 3 == shift)]
-      
+      strandStops <- geneRightPos[strandOnly]
     } else {
+      ## Get and set up the vectors to work through the current reverse frame.
       strandID <- "-"
       cFrame <- frameID - 3
       shift <- cFrame %% 3
@@ -529,57 +549,178 @@ AssessGenes <- function(geneLeftPos,
       
       strandOnly <- which(geneStrand == strandID)
       
-      strandStarts <- genomeLength - geneRightPos[strandOnly] + 1
-      names(strandStarts) <- as.character(strandOnly)
-      
-      strandFrameStarts <- strandStarts[(strandStarts %% 3 == shift)]
+      strandStarts <- genomeLength - geneRightPos[strandOnly] + 1L
+      strandStops <- genomeLength - geneLeftPos[strandOnly] + 1L
     }
     
-    ## In order to handle genes at the boundary of the genome, two additional
-    ## in-frame positions need to be added to the stops vector for the frame:
-    ## the start of the first possible in-frame codon and the start of the
-    ## last possible in-frame codon.
+    ## Add in names to the strand-specific gene position vectors so that the
+    ## categorization of genes can be mapped back to the order the genes were
+    ## inputted into the function.
+    names(strandStarts) <- names(strandStops) <- as.character(strandOnly)
+   
+    ## Get and set up the predicted gene position vectors for the current frame. 
+    strandFrameStarts <- strandStarts[(strandStarts %% 3 == shift)]
+    strandFrameStops <- strandStops[(strandStarts %% 3 == shift)]
     
+    ## In order to handle genes at the boundary of the genome, additional
+    ## in-frame positions need to be added to the stops vector for the frame:
+    ## the start of the first possible in-frame codon and the start(s) of the
+    ## last one or two possible in-frame codon. The start of the last codon is
+    ## always included. The start of the second to last possible codon is only
+    ## included if the last codon is not completely within the genome.a
     finalCodonStart <- ifelse(genomeLength %% 3 == shift, genomeLength,
                               ifelse((genomeLength - 1L) %% 3 == shift,
                                      genomeLength - 1L, genomeLength - 2L))
     
-    frameStops <- c(cFrame, stops[[frameID]], finalCodonStart)
+    if ((finalCodonStart + 2L) == genomeLength) {
+      frameStops <- sort(unique(c(cFrame, stops[[frameID]],
+                                  finalCodonStart)))
+    } else {
+      frameStops <- sort(unique(c(cFrame, stops[[frameID]],
+                                  finalCodonStart - 3L, finalCodonStart)))
+    }
     
+    ## If a predicted gene has a stop that is beyond the end of region, it will
+    ## be necessary to skip subsequent genome stops until the predicted stop.
+    skipStop <- logical(length(frameStops))
+    
+    ## For each frame, loop through each stop-to-stop region and categorize the
+    ## gene(s) or proteomics hits within that region if there are any.
     for (idx in seq(1, length(frameStops) - 1)) {
-      ## Set boundaries of the region to examine.
-      orfStart <- frameStops[idx]
-      orfEnd <- frameStops[idx + 1]
-      orfRange <- checkRange(c(orfStart, orfEnd))
+      ## Skip the current region based on the region-ending stop. See below and
+      ## the checkNextStops() function for how region skipping is determined.
+      if (skipStop[idx + 1]) {
+        next()
+      }
       
-      ## For each ORF, start by assuming there are no valid conserved starts,
-      ## and no proteomics hits mapping to that ORF.
+      ## Set boundaries of the region to examine.
+      ## The first variable corresponds to the region-starting stop
+      ## The second variable corresponds to the region-ending stop.
+      ## The third variable is the range between those two positions.
+      regStart <- frameStops[idx]
+      regEnd <- frameStops[idx + 1]
+      regRange <- regStart:regEnd
+      
+      ## For each region, start by assuming there are no valid conserved
+      ## starts, and no proteomics hits mapping to that region.
       noConStarts <- TRUE
       noProt <- TRUE
       
-      ## Get the gene start(s) in that region.
-      regionalPredictedStart <-strandFrameStarts[((strandFrameStarts >= orfStart) &
-                                                    (strandFrameStarts <= orfEnd))]
+      ## Get the gene start(s) and gene stop(s) in that region.
+      regionalPredictedStart <- strandFrameStarts[((strandFrameStarts >= regStart) &
+                                                     (strandFrameStarts <= regEnd))]
+      regionalPredictedStop <- strandFrameStops[((strandFrameStarts >= regStart) &
+                                                   (strandFrameStarts <= regEnd))]
+      
+      ## Get the index for the gene(s).
+      ## This is based on the order the genes were inputted into the function.
       geneIdx <- as.integer(names(regionalPredictedStart))
       
+      ## If there is more than one gene in the region, warn the user and
+      ## put the genes into the no evidence category.
       if (length(regionalPredictedStart) > 1) {
-        warning("More than one in-frame gene start between stops ",
-                orfStart, " and ", orfEnd, " in frame ", frameID, ".",
-                paste(" A gene start is at ", regionalPredictedStart,
-                      " (gene ", geneIdx, ").", sep = ""))
+        ## Get the forward strand positions for both the region
+        ## and the genes to use in the warning message.
+        if (frameID <= 3) {
+          fwdRegStart <- regStart
+          fwdRegEnd <- regEnd
+          fwdPredStart <- regionalPredictedStart
+          fwdPredStop <- regionalPredictedStop
+        } else {
+          fwdRegStart <- genomeLength - regStart + 1L
+          fwdRegEnd <- genomeLength - regEnd + 1L
+          fwdPredStart <- genomeLength - regionalPredictedStart + 1L
+          fwdPredStop <- genomeLength - regionalPredictedStop + 1L
+        }
         
+        ## Warn the user about multiple genes within a particular region.
+        warning("More than one in-frame gene start between stops ",
+                fwdRegStart, " and ", fwdRegEnd, " in frame ", frameID, ".",
+                paste0(" A gene start is at ", fwdPredStart, " with stop at ",
+                       fwdPredStop, " (gene ", geneIdx, ")."),
+                " These genes will be placed into the no evidence category.")
+        
+        lastPredStop <- regionalPredictedStop[length(regionalPredictedStart)]
+        
+        ## If the most downstream predicted stop in the set of genes is
+        ## downstream of the region-ending stop, skip subsequent genome
+        ## stops as long as no predicted genes will be skipped.
+        if ((lastPredStop - 2L) > regEnd) {
+          skipStop[checkNextStops(lastPredStop)] <- TRUE
+        }
+        
+        ## Assign those genes to the no evidence category.
         catGeneAssignment[geneIdx] <- "Y CS- PE-"
         next()
       }
       
+      ## If there is one gene in the region, make sure the stop makes sense.
+      ## Note that the region-ending stop position corresponds to the first
+      ## position of the stop codon while the predicted stop position should
+      ## correspond to the last position of the stop codon.
+      if (length(regionalPredictedStart) == 1) {
+        ## Check if the predicted stop is beyond the end of the region.
+        beyondEnd <- ((regionalPredictedStop - 2L) > regEnd)
+        
+        ## Check if the predicted stop is out of frame.
+        outOfFrame <- ((regionalPredictedStop - 2L) %% 3 != shift)
+        
+        ## Check if the predicted stop is the before the end of the region.
+        beforeEnd <- ((regionalPredictedStop - 2L) < regEnd)
+        
+        ## If the predicted stop is not correct for the gene, warn the user
+        ## appropriately and place the gene in the no evidence category.
+        if (beyondEnd || outOfFrame || beforeEnd) {
+          ## Get the forward strand positions for both the correct stop
+          ## and the predicted stop to use in the warning message(s).
+          if (frameID <= 3) {
+            fwdRightStop <- regEnd + 2L
+            fwdPredStop <- regionalPredictedStop
+          } else {
+            fwdRightStop <- genomeLength - regEnd + 1L
+            fwdPredStop <- genomeLength - regionalPredictedStop + 1L
+          }
+          
+          ## Print the warning message.
+          ## Skip subsequent genome stops if necessary.
+          if (beyondEnd) {
+            warning("The predicted stop for gene ", geneIdx, ", position ", fwdPredStop,
+                    ", is downstream of the correct stop. The stop should be at ",
+                    fwdRightStop, ". The gene is in frame ", frameID, ".",
+                    " The gene will be placed into the no evidence category.")
+            
+            ## If the predicted stop is downstream of the end of the region, skip
+            ## subsequent genome stops between the correct stop and the predicted
+            ## one as long as no predicted genes will be skipped.
+            skipStop[checkNextStops(regionalPredictedStop)] <- TRUE
+            
+          } else if (outOfFrame) {
+            warning("The predicted stop for gene ", geneIdx, ", position ",
+                    fwdPredStop, ", is out of frame. The stop should be at ",
+                    fwdRightStop, ". The gene is in frame ", frameID, ".",
+                    " The gene will be placed into the no evidence category.")
+            
+          } else if (beforeEnd) {
+            warning("The predicted stop for gene ", geneIdx, ", position ", fwdPredStop,
+                    ", is upstream of the correct stop. The stop should be at ",
+                    fwdRightStop, ". The gene is in frame ", frameID, ".",
+                    " The gene will be placed into the no evidence category.")
+          }
+          
+          ## Assign the gene to the no evidence category.
+          catGeneAssignment[geneIdx] <- "Y CS- PE-"
+          next()
+        }
+      }
+      
       if (useProt) {
         ## Get the protein hits in that region.
-        regionalProteinHits <- frameProteinHits[orfRange] > 0
+        regionalProteinHits <- frameProteinHits[regRange] > 0
         regionalProteinHitStarts <- which(diff(c(0, regionalProteinHits)) == 1) # start
         
         ## If there are protein hits, determine the start of the first protein hit.
         if (length(regionalProteinHitStarts) >= 1) {
-          firstProteinHitStart <- min(orfRange[regionalProteinHitStarts])
+          firstProteinHitStart <- min(regRange[regionalProteinHitStarts])
           
           if ((isNTerm) && (length(regionalPredictedStart) > 0)) {
             protAligned <- (firstProteinHitStart == regionalPredictedStart) ||
@@ -597,13 +738,13 @@ AssessGenes <- function(geneLeftPos,
         if ((length(regionalPredictedStart) <= 0) ||
             !(is.na(strandConservedStarts[regionalPredictedStart]))) {
           ## .... find which start positions have some conservation and coverage in the region.
-          allConStartIdxs <- which((strandConservedStarts[orfRange] / strandCoverage[orfRange] > 0) &
-                                     (strandCoverage[orfRange] >= minCov) &
-                                     (((orfRange - cFrame) %% 3) == 0))
+          allConStartIdxs <- which((strandConservedStarts[regRange] / strandCoverage[regRange] > 0) &
+                                     (strandCoverage[regRange] >= minCov) &
+                                     (((regRange - cFrame) %% 3) == 0))
           
           if (length(allConStartIdxs) >= 1) {
             ## If there are any such positions, get all possible conserved starts.
-            allConservedStarts <- orfRange[allConStartIdxs]
+            allConservedStarts <- regRange[allConStartIdxs]
             
             ## Only the conserved starts upstream of the protein hits can be used as evidence.
             if (!noProt) {
@@ -629,7 +770,8 @@ AssessGenes <- function(geneLeftPos,
         }
       }
       
-      ## Skip the current ORF if there is no proteomics hit or gene start mapping to it.
+      ## Skip the current region if there is no proteomics hit
+      ## or gene start mapping to it.
       if ((noProt) && (length(regionalPredictedStart) <= 0)) {
         next()
       }
@@ -638,12 +780,12 @@ AssessGenes <- function(geneLeftPos,
       ## in that region. Requires a predicted gene in that region.
       if ((length(regionalPredictedStart) > 0) && (useCons)) {
         ## Find which positions have some coverage and stop codon conservation in the region.
-        allConStopIdxs <- which((strandConservedStops[orfRange] / strandCoverage[orfRange] > minConCovRatio_Stop) &
-                                  (strandCoverage[orfRange] >= minCov) &
-                                  (((orfRange - cFrame) %% 3) == 0))
+        allConStopIdxs <- which((strandConservedStops[regRange] / strandCoverage[regRange] > minConCovRatio_Stop) &
+                                  (strandCoverage[regRange] >= minCov) &
+                                  (((regRange - cFrame) %% 3) == 0))
         
         if (length(allConStopIdxs) >= 1) {
-          allConservedStops <- orfRange[allConStopIdxs]
+          allConservedStops <- regRange[allConStopIdxs]
           
           ## Only the conserved stops upstream of the protein hits can be used as evidence.
           if (!noProt) {
@@ -652,7 +794,7 @@ AssessGenes <- function(geneLeftPos,
             currConStops <- allConservedStops
           }
           
-          geneLen <- (orfEnd + 2L) - (regionalPredictedStart) + 1L
+          geneLen <- (regEnd + 2L) - (regionalPredictedStart) + 1L
           
           ## Only the conserved stops early on in the gene can be used as evidence.
           condition1 <- (currConStops >= regionalPredictedStart)
@@ -671,7 +813,7 @@ AssessGenes <- function(geneLeftPos,
             ## If there is no protein evidence and either there are (multiple) valid conserved stops or there is
             ## a strong conserved start downstream of the stops, put the gene in the conserved stop category.
             if ((noProt) && (multiStops || conStopConStart)) {
-              ## At least one Prodigal start in that region
+              ## At least one predicted start in that region
               ## Conserved stops disprove the gene --> "Y CS! PE-"
               catGeneAssignment[geneIdx] <- "Y CS! PE-"
               next()
@@ -680,7 +822,7 @@ AssessGenes <- function(geneLeftPos,
             ## If there is protein evidence and there is a strong conserved start downstream of the stops,
             ## put the gene in the conserved stop category.
             if ((!noProt) && conStopConStart) {
-              ## At least one Prodigal start in that region
+              ## At least one predicted start in that region
               ## Conserved stops disprove the gene --> "Y CS! PE+"
               catGeneAssignment[geneIdx] <- "Y CS! PE+"
               next()
@@ -689,9 +831,9 @@ AssessGenes <- function(geneLeftPos,
         }
       }
       
-      ## If there is only a gene start mapping to the current ORF...
+      ## If there is only a gene start mapping to the current region...
       if ((noProt) && (noConStarts) && (length(regionalPredictedStart) > 0)) {
-        ## At least one Prodigal start in that region
+        ## At least one predicted start in that region
         ## No protein hits and no conserved starts --> Category 7, "Y CS- PE-"
         catGeneAssignment[geneIdx] <- "Y CS- PE-"
         next()
@@ -700,39 +842,39 @@ AssessGenes <- function(geneLeftPos,
       ## Check if there are any protein hits within that region (section for categories 1-3).
       if (!noProt) {
         ## There are protein hits in that region.
-        ## Check if there is any Prodigal start within that region.
+        ## Check if there is any predicted start within that region.
         if (length(regionalPredictedStart) <= 0) {
-          ## No prodigal starts within that region.
+          ## No predicted starts within that region.
           
-          if ((orfEnd - orfStart) >= minMissORFLen) {
-            ## Check if ORF is nested, if necessary.
+          if ((regEnd - regStart) >= minMissORFLen) {
+            ## Check if region is nested, if necessary.
             if ((allowNestedORFs) || !(isNestedORF())) {
               if (frameID <= 3) {
-                cat1Pos <- c(orfStart + 3L, orfEnd + 2L)
+                cat1Pos <- c(regStart + 3L, regEnd + 2L)
               } else {
-                cat1Pos <- genomeLength - c(orfStart + 3L, orfEnd + 2L) + 1L
+                cat1Pos <- genomeLength - c(regStart + 3L, regEnd + 2L) + 1L
               }
               
               ## Check if there is any conserved start (upstream of the first protein hit) in that region.
               if (noConStarts) {
-                ## Protein hit with no conserved start upstream and no Prodigal start --> Category 1A
+                ## Protein hit with no conserved start upstream and no predicted start --> Category 1A
                 ## "N CS- PE+"
                 cat1A_Counter <- cat1A_Counter + 1L
                 
-                cat1A_ORFs[cat1A_Counter, ] <- c(cat1Pos, orfEnd - orfStart, frameID, checkOtherFrames())
+                cat1A_ORFs[cat1A_Counter, ] <- c(cat1Pos, regEnd - regStart, frameID, checkOtherFrames())
               } else {
-                ## Protein hit with conserved start upstream but no Prodigal start --> Category 1B
+                ## Protein hit with conserved start upstream but no predicted start --> Category 1B
                 ## "N CS< PE+"
                 cat1B_Counter <- cat1B_Counter + 1L
                 
-                cat1B_ORFs[cat1B_Counter, ] <- c(cat1Pos, orfEnd - orfStart, frameID, checkOtherFrames())
+                cat1B_ORFs[cat1B_Counter, ] <- c(cat1Pos, regEnd - regStart, frameID, checkOtherFrames())
               }
             }
           }
           
           next()
         } else {
-          ## At least one Prodigal start in that region
+          ## At least one predicted start in that region
           
           ## Check if there is any conserved start upstream of the first protein hit in that region.
           if (noConStarts) {
@@ -794,8 +936,9 @@ AssessGenes <- function(geneLeftPos,
         next()
       }
       
-      ## There is at least one good conserved start somewhere in the ORF but it is not aligned with given start.
-      ## If there is a bad conserved start aligned with the given start, ...
+      ## There is at least one good conserved start somewhere in the region but
+      ## it is not aligned with given start. If there is a bad conserved start
+      ## aligned with the given start, ...
       if ((length(badConStartIdxs) >= 1) && (any(conStarts[badConStartIdxs] == regionalPredictedStart))) {
         ## ... check where the best conserved starts are relative to the given gene start.
         pickDownstream <- (all(conStarts[bestConStartIdxs] > regionalPredictedStart))