+++ R/MapAssessmentData.R
@@ -64,7 +64,7 @@
 #' related genomes have been sorted from most distantly related to most closely related in relation to the central genome.
 #' Default value is 1000.
 #' 
-#' @param startCodons A charcter vector consisting of three-letter DNA strings to use as the start codons when finding
+#' @param startCodons A character vector consisting of three-letter DNA strings to use as the start codons when finding
 #' evolutionarily conserved starts.
 #' 
 #' @param ema_AlphaVal The alpha value to use when calculating the exponential moving average over an alignment derived
@@ -101,7 +101,7 @@
 #' will be kept and the rest of the hits will be dropped. If all the proteomics hits have the same score or if \code{protHits_Threshold}
 #' is zero, no thresholding will occur and no hits will be dropped.
 #' 
-#' Please note that \code{protHits_IsNTerm} has no affect on how the proteomics evidence is mapped to the central genome but it can be
+#' Please note that \code{protHits_IsNTerm} has no effect on how the proteomics evidence is mapped to the central genome but it can be
 #' used to affect how genes are assessed and categorized in \code{\link{AssessGenes}}.
 #'
 #' Evolutionarily conserved starts and conserved stop are found by first measuring how far the related genomes are from the central
@@ -809,7 +809,7 @@ MapAssessmentData <- function(genomes_DBFile,
         
         ## Add to the coverage.
         fwdCov[covRange] <- fwdCov[covRange] + 1L
-        revCov[genomeLen - covRange + 1] <- revCov[genomeLen - covRange + 1] + 1L
+        revCov[genomeLen - covRange - 1] <- revCov[genomeLen - covRange - 1] + 1L
         
         ## --------------------------------------------------------------------------- ##